Feature: Initiate page changes via an R function
This commit actually does a few things:

1. Adds an R function change_page(), to allow the current
page to be changed via R. This is useful for interactive
click handlers, for example.

2. Adds convenience functions for R application code to
easily see what the current page is, any query params on
it, and to ensure they only run on a particular page.

3. Removes the JS-side validation of routes, and moves
route validation entirely to the R side.

4. Removes the dependency on the "shinyjs" package (which
was only used to set up the JS-side route validation).