New(ish) core Shiny functionality makes the JS unnecessary!
Earlier this year Shiny added some functionality, so that
shiny::getUrlHash() updates reactively whenever the page's
URL fragment changes. In conjunction with shiny::updateQueryString(),
this means that we no longer need any custom JS at all, in order
to be notified of updates to the fragment, and to do an R-initiated
update of the fragment!

Scenario 1: User clicks on link in the webpage
This triggers a browser "hashchange" event. Shiny's JS code
has an event handler watching for that, notices, and reactively
updates all consumers of shiny::getUrlHash().

Scenario 2: R code needs to initiate a page change
Can use shiny::updateQueryString(). Despite the name, if the string
passed in starts with "#", it will update the fragment rather than
the query string proper! It will also then trigger hashchange,
sending the updated page back to all consumers of shiny::getUrlHash().

The only scenario we can't cover, is if some JS on the page changes
the URL via history.pushState(). But page.js couldn't deal with
that either, because there's no event handler for it.