Use history.replaceState() for some URL changes
If the URL changes due to a redirect away from
an invalid URL, that's not strictly "navigation",
and ideally shouldn't put a new entry in the browser
history.

We can use the "replace" param for Shiny's API to
achieve this, but unfortunately Shiny doesn't
update clientData$url_hash when you do that. So,
we have to hack window.history.replaceState() to
do so.